backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
redis==5.1.1
celery==5.4.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
sendgrid==6.11.0
twilio==9.3.6
python-dotenv==1.0.1
stripe==7.7.0
email-validator==2.2.0
boto3==1.34.0
Pillow==11.0.0
jinja2==3.1.4